Extract common docs to Hex.Package
